tests/test_api.py
"""
tests/test_api.py — Integration and unit tests for CodeLens.
Run with: pytest tests/ -v
"""

import io
import os
import sys
import zipfile
import pytest

sys.path.insert(0, os.path.dirname(os.path.dirname(__file__)))
os.environ.setdefault("GEMINI_API_KEY", "test_key_placeholder")

from fastapi.testclient import TestClient
from backend.main import app

client = TestClient(app)


# ── Health ─────────────────────────────────────────────────────────────────────
def test_health_structure():
    resp = client.get("/api/health")
    assert resp.status_code == 200
    data = resp.json()
    assert "status" in data
    assert "services" in data
    assert "backend" in data["services"]
    assert "database" in data["services"]
    assert "llm" in data["services"]
    assert data["services"]["backend"]["status"] == "ok"

def test_health_latency():
    data = client.get("/api/health").json()
    assert "latency_ms" in data
    assert isinstance(data["latency_ms"], (int, float))


# ── Pages ──────────────────────────────────────────────────────────────────────
def test_home_page():
    resp = client.get("/")
    assert resp.status_code == 200
    assert "text/html" in resp.headers["content-type"]

def test_status_page():
    resp = client.get("/status")
    assert resp.status_code == 200
    assert "text/html" in resp.headers["content-type"]


# ── Upload validation ──────────────────────────────────────────────────────────
def test_upload_non_zip_rejected():
    resp = client.post("/api/upload",
                       files={"file": ("test.txt", b"not a zip", "text/plain")})
    assert resp.status_code == 400
    assert "zip" in resp.json()["detail"].lower()

def test_upload_invalid_zip_rejected():
    resp = client.post("/api/upload",
                       files={"file": ("test.zip", b"not a zip at all", "application/zip")})
    assert resp.status_code == 400

def test_upload_empty_filename():
    resp = client.post("/api/upload",
                       files={"file": ("", b"data", "application/zip")})
    assert resp.status_code in (400, 422)


# ── GitHub validation ──────────────────────────────────────────────────────────
def test_github_empty_url():
    resp = client.post("/api/github", json={"repo_url": ""})
    assert resp.status_code == 400

def test_github_invalid_url():
    resp = client.post("/api/github", json={"repo_url": "not-a-url"})
    assert resp.status_code in (400, 422, 500)


# ── Ask validation ─────────────────────────────────────────────────────────────
def test_ask_empty_question():
    resp = client.post("/api/ask", json={
        "session_id": "fake-session",
        "question": "",
    })
    assert resp.status_code == 400

def test_ask_missing_session():
    resp = client.post("/api/ask", json={
        "session_id": "00000000-0000-0000-0000-000000000000",
        "question": "Where is auth handled?",
    })
    assert resp.status_code == 404


# ── Session ────────────────────────────────────────────────────────────────────
def test_get_nonexistent_session():
    resp = client.get("/api/sessions/nonexistent-id")
    assert resp.status_code == 404

def test_history_returns_list():
    resp = client.get("/api/history/fake-session-id")
    assert resp.status_code == 200
    assert isinstance(resp.json()["history"], list)

def test_export_nonexistent_session():
    resp = client.get("/api/export/nonexistent-session")
    assert resp.status_code == 404


# ── Indexer unit tests ─────────────────────────────────────────────────────────
def test_indexer_chunking():
    """Chunker produces correct chunks with line numbers."""
    import tempfile
    from pathlib import Path
    from backend.indexer import CodebaseIndexer

    idx = CodebaseIndexer.__new__(CodebaseIndexer)
    lines = [f"x_{i} = {i}" for i in range(100)]
    content = "\n".join(lines)
    chunks = idx._chunk_file("test/file.py", content)

    assert len(chunks) > 1
    assert chunks[0]["file"] == "test/file.py"
    assert chunks[0]["line_start"] == 1
    assert all("raw" in c for c in chunks)


def test_indexer_skips_no_code_files():
    """ZIP with only binary files should raise ValueError."""
    import tempfile
    from pathlib import Path
    from backend.indexer import CodebaseIndexer

    with tempfile.TemporaryDirectory() as tmpdir:
        (Path(tmpdir) / "binary.bin").write_bytes(b"\x00\x01\x02\x03")
        idx_dir = Path(tmpdir) / "index"
        idx = CodebaseIndexer(str(idx_dir))
        with pytest.raises(ValueError, match="No indexable"):
            idx.index_directory("test-session", tmpdir)


def test_indexer_windows_path_separator():
    """File paths should always use forward slashes."""
    from backend.indexer import CodebaseIndexer
    idx = CodebaseIndexer.__new__(CodebaseIndexer)
    chunks = idx._chunk_file("src\\auth\\login.py", "def login(): pass")
    # Should NOT contain backslash — but we pass already-fixed paths
    # This tests the chunk structure
    assert "file" in chunks[0]
    assert "line_start" in chunks[0]
    assert "raw" in chunks[0]