requirements.txt
# ── Web framework ──────────────────────────────────────────────────────────────
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
httpx==0.27.0

# ── Google Gemini — LLM (gemini-1.5-flash) + Embeddings (embedding-001) ────────
# Works on Python 3.13 Windows without any C compiler.
google-generativeai==0.8.3

# ── Numeric — for cosine similarity search (replaces FAISS) ────────────────────
numpy>=2.0.0

# ── Utilities ──────────────────────────────────────────────────────────────────
python-dotenv==1.0.1